requirements.txt
chromadb==1.0.7
fastapi==0.115.9
motor==3.7.0
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.12.1
python-dotenv==1.1.0
python-multipart==0.0.20
uvicorn==0.34.2